pyproject.toml
[project]
name = "smart-cooling"
version = "0.1.0"
description = "Smart Cooling integration for Home Assistant with learning capabilities"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Mark Aggar", email = "mark@aggar.com"}
]
requires-python = ">=3.11"
keywords = ["home-assistant", "smart-home", "hvac", "cooling", "climate"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = []

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-homeassistant-custom-component>=0.13.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
]

[project.urls]
Homepage = "https://github.com/markaggar/smart-cooling"
Repository = "https://github.com/markaggar/smart-cooling"
Issues = "https://github.com/markaggar/smart-cooling/issues"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["custom_components*"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "S",   # flake8-bandit
]
ignore = [
    "S101",  # Use of assert (needed for tests)
]

[tool.ruff.per-file-ignores]
"tests/**/*.py" = ["S101"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["custom_components/smart_cooling"]
omit = ["tests/*"]
